backend/requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
PyJWT
requests
google-cloud-speech
openai
langchain
pinecone-client